add action to toggle quick documentation translation
