+++ src/main/kotlin/cn/yiiguxing/plugin/translate/documentation/TranslateDocumentationTask.kt
@@ -1,0 +1,42 @@
+package cn.yiiguxing.plugin.translate.documentation
+
+import cn.yiiguxing.plugin.translate.action.TranslateDocumentationAction
+import com.intellij.openapi.progress.ProgressManager
+import org.jetbrains.concurrency.runAsync
+import java.util.concurrent.TimeoutException
+
+class TranslateDocumentationTask(val text: String) {
+
+    private val totalTimeToWaitMs = 3_000
+    private val timeToBlockMs = 100
+
+    private var tries = totalTimeToWaitMs / timeToBlockMs
+
+    //execute on a different thread outside read action
+    private val promise = runAsync {
+        getTranslatedDocumentation(text)
+    }
+
+    fun onSuccess(callback: (String) -> Unit) {
+        promise.onSuccess(callback)
+    }
+
+    fun nonBlockingGet(): String? {
+        //blocking for the whole time can lead to ui freezes, so we need to periodically do `checkCanceled`
+        while (tries > 0) {
+            tries -= 1
+            ProgressManager.checkCanceled()
+            try {
+                return promise.blockingGet(timeToBlockMs)
+            } catch (t: TimeoutException) {
+                //ignore
+            } catch (e: Throwable) {
+                TranslateDocumentationAction.logAndShowWarning(e, null)
+                return null
+            }
+        }
+
+        //translation is not ready yet, show original documentation
+        return null
+    }
+}
