+++ src/main/kotlin/cn/yiiguxing/plugin/translate/provider/TranslatingDocumentationProvider.kt
@@ -1,13 +1,10 @@
 package cn.yiiguxing.plugin.translate.provider
 
 import cn.yiiguxing.plugin.translate.Settings
-import cn.yiiguxing.plugin.translate.action.TranslateDocumentationAction
+import cn.yiiguxing.plugin.translate.documentation.TranslateDocumentationTask
 import com.intellij.codeInsight.documentation.DocumentationManager
 import com.intellij.lang.documentation.DocumentationProviderEx
-import com.intellij.openapi.progress.ProgressManager
 import com.intellij.psi.PsiElement
-import org.jetbrains.concurrency.runAsync
-import java.util.concurrent.TimeoutException
 
 /**
  * Translates documentation computed by another documentation provider. It should have
@@ -51,7 +48,7 @@ class TranslatingDocumentationProvider : DocumentationProviderEx() {
         }
 
         //to reuse long running translation task
-        private var lastTranslation: TranslationTask? = null
+        private var lastTranslation: TranslateDocumentationTask? = null
 
         private fun translate(text: String?): String? {
             if (text.isNullOrEmpty()) return null
@@ -60,42 +57,11 @@ class TranslatingDocumentationProvider : DocumentationProviderEx() {
 
             val task =
                 if (lastTask != null && lastTask.text == text) lastTask
-                else TranslationTask(text)
+                else TranslateDocumentationTask(text)
 
             lastTranslation = task
 
             return task.nonBlockingGet()
-        }
-
-        private data class TranslationTask(val text: String) {
-            private val totalTimeToWaitMs = 3_000
-            private val timeToBlockMs = 100
-
-            private var tries = totalTimeToWaitMs / timeToBlockMs
-
-            //execute on a different thread outside read action
-            private val promise = runAsync {
-                TranslateDocumentationAction.getTranslatedDocumentation(text)
-            }
-
-            fun nonBlockingGet(): String? {
-                //blocking for the whole time can lead to ui freezes, so we need to periodically do `checkCanceled`
-                while (tries > 0) {
-                    tries -= 1
-                    ProgressManager.checkCanceled()
-                    try {
-                        return promise.blockingGet(timeToBlockMs)
-                    } catch (t: TimeoutException) {
-                        //ignore
-                    } catch (e: Throwable) {
-                        TranslateDocumentationAction.logAndShowWarning(e, null)
-                        return null
-                    }
-                }
-
-                //translation is not ready yet, show original documentation
-                return null
-            }
         }
     }
 }
